Remove extraneous code and references
